+++ src/mol-model/structure/structure/structure.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2017-2023 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2017-2024 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author David Sehnal <david.sehnal@gmail.com>
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
@@ -158,6 +158,14 @@ class Structure {
      */
     get isCoarseGrained() {
         return this.models.some(m => Model.isCoarseGrained(m));
+    }
+
+    /**
+     * True if all models provide explicit bond information (i.e., `chem_comp_bond` is defined).
+     * @see Model.hasChemCompBond
+     */
+    get hasChemCompBond() {
+        return this.models.length && this.models.every(m => Model.hasChemCompBond(m));
     }
 
     get isEmpty() {